move browser test to its own file inside SupportsWireables
